+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/KubernetesClient.java
@@ -73,9 +73,9 @@ import io.fabric8.kubernetes.api.model.coordination.v1.DoneableLease;
 import io.fabric8.kubernetes.api.model.coordination.v1.Lease;
 import io.fabric8.kubernetes.api.model.coordination.v1.LeaseList;
 import io.fabric8.kubernetes.client.dsl.*;
-import io.fabric8.kubernetes.api.model.apiextensions.CustomResourceDefinition;
-import io.fabric8.kubernetes.api.model.apiextensions.CustomResourceDefinitionList;
-import io.fabric8.kubernetes.api.model.apiextensions.DoneableCustomResourceDefinition;
+import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.CustomResourceDefinition;
+import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.CustomResourceDefinitionList;
+import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.DoneableCustomResourceDefinition;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.internal.RawCustomResourceOperationsImpl;
 import io.fabric8.kubernetes.client.extended.leaderelection.LeaderElectorBuilder;
@@ -97,6 +97,14 @@ public interface KubernetesClient extends Client {
    * @return NonNamespaceOperation object for CustomResourceDefinition
    */
   NonNamespaceOperation<CustomResourceDefinition, CustomResourceDefinitionList, DoneableCustomResourceDefinition, Resource<CustomResourceDefinition, DoneableCustomResourceDefinition>> customResourceDefinitions();
+
+  /**
+   * API entrypoint for apiextensions resources. Currently support both
+   * v1 and v1beta1
+   *
+   * @return ApiextensionsAPIGroupDSL which routes to v1 or v1beta1
+   */
+  ApiextensionsAPIGroupDSL apiextensions();
 
   /**
    * Typed API for managing CustomResources. You would need to provide POJOs for